+++ CHANGELOG.md
@@ -15,4 +15,5 @@ if the major version hasn't changed.
 
 ### Added
 
+- Add support for pinned views (#336)
 - Initial open-source release of `fiberplane-rs`.